\docType{package}
\name{FREGAT-package}
\alias{FREGAT-package}
\title{FREGAT: Family REGional Association Tests}
\description{
	The FREGAT package supplies the most common and efficient
	tests for the region-based association analysis aimed at identification
	of rare genetic variants for family-based, genetically related
	or population samples. Tests implemented are FFBSKAT (fast family-based
	sequence kernel association test), FFBSKAT optimal, famBT (family burden
	test), famFLM (family functional linear model, a functional data analysis
	approach), and MLR (standard multiple linear regression). The methods
	provide regional association testing of a set of SNPs with a continuous
	phenotype in the presence of additional covariates and within-family
	correlations.
}
\details{
Package: FREGAT\cr
Type: Package\cr
License: GPLv3
}
\author{
	Nadezhda Belonogova <belon@bionet.nsc.ru>\cr
	Gulnara Svishcheva <gulsvi@mail.ru>\cr\cr
	The authors are very grateful to Dr. Anatoly Kirichenko for technical support\cr
	and Prof. Tatiana Axenovich for scientific guidance.\cr
}
\references{
	Belonogova N.M., Svishcheva G.R., Axenovich T.I. (2016) FREGAT: an R package for region-based association analysis. Bioinformatics,  V32(15), P. 2392-2393.\cr
}
