\docType{package}
\name{FREGAT-package}
\alias{FREGAT-package}
\title{FREGAT: Family REGional Association Tests}
\description{
	The FREGAT package supplies the most common and efficient
	tests for the region-based association analysis aimed at identification
	of rare genetic variants for family-based, genetically related
	or population samples. Tests implemented are FFBSKAT (fast family-based
	sequence kernel association test), FFBSKAT optimal, famBT (family burden
	test), famFLM (family functional linear model, a functional data analysis
	approach), and MLR (standard multiple linear regression). The methods
	provide regional association testing of a set of SNPs with a continuous
	phenotype in the presence of additional covariates and within-family
	correlations.
}
\details{
Package: FREGAT\cr
Type: Package\cr
Version: 1.0\cr
License: GPLv3
}
\author{
	Nadezhda Belonogova <belon@bionet.nsc.ru>\cr
	Gulnara Svishcheva <gulsvi@mail.ru>\cr\cr
	The authors are very grateful to Dr. Anatoly Kirichenko for technical support\cr
	and Prof. Tatiana Axenovich for scientific guidance.\cr
}
\references{
	Svishcheva G.R., Belonogova N.M. and Axenovich T.I. (2014) FFBSKAT: Fast Family-Based Sequence Kernel Association Test. PLoS ONE 9(6): e99407.\cr\cr
	Svishcheva G.R., Belonogova N.M. and Axenovich T.I. (2015) Region-based association test for familial data under functional linear models. PLoS ONE 10(6): e0128999.\cr
}
